classification options minor code refactor for simplicity
I use the $index for the option value and just look that up in the
options rather than having a second object keyed off the ID with the
same values.

This also fixes some of the visual layout treatments for spacing, etc.

Also, the “verify” button switches to “save” if you have chosen
anything other than the default “best” option.