Bug 1233653 - Use start time not submit time in running_time_remaining()
Since the average duration (unfortunately named `running_eta` in the API
response) is from the start timestamp (ie the time the job actually
started), not the time the job was scheduled - since the job may have
been waiting to start for some time.

The function's response is also now inverted: if it's overdue the
integer returned is negative, when previously it would have been
positive.

Since `running_time_remaining()` now uses start_timestamp, we must only
call if for running jobs, which is not a problem since the values are
not used for other states anyway (even though we tried to calculate it
for all jobs previously).